update cookie modal for all pages
