+++ src/pages/_app.tsx
@@ -14,6 +14,7 @@ import 'aos/dist/aos.css';
 import { useEffect, useState } from 'react';
 import Script from 'next/script';
 import Head from 'next/head';
+import CookieModal from '@/components/CookieModal';
 
 config.autoAddCss = false;
 
@@ -85,6 +86,7 @@ export default function App({ Component, pageProps }: AppProps) {
       />
         </>
       )}
+      <CookieModal />
       <Component {...pageProps} />
     </>
   );
