+++ src/pages/index.tsx
@@ -10,26 +10,12 @@ import Services from "@/components/Service/Service";
 import Testimonials from "@/components/Testimonials/Testimonials";
 import TryOpenMetadata from "@/components/TryOpenMetadata/TryOpenMetadata";
 import Head from "next/head";
-import Link from "next/link";
-import { useCallback, useEffect, useState } from "react";
 
 const Development = () => {
-  const [cookieValue, setCookieValue] = useState<string | null>(null)
-
-    const handleButtonClick = useCallback((choice: string) => {
-        localStorage.setItem('userCookie', choice)
-        setCookieValue(choice)
-    }, [])
-
   const handleTryOpenMetadataClick = () => {
     const element = document.querySelector("#try-openmetadata");
     element?.scrollIntoView({ behavior: "smooth", block: "end" })
   }
-
-  useEffect(() => {
-    const userCookie = localStorage.getItem('userCookie')
-    setCookieValue(userCookie)
-}, [])
 
   return (
     <>
@@ -91,31 +77,6 @@ const Development = () => {
           <NavbarDev onClick={handleTryOpenMetadataClick} />
           <NavbarStrip />
         </div>
-        {!cookieValue && (
-                <div className="fixed bottom-0 bg-[#F8FBFC] z-[20] p-6 shadow-shadow card sm:m-5 sm:max-w-xl lg:max-w-3xl">
-                    <p className="text-[#292929] font-light">
-                        By clicking &quot;Accept&quot;, you agree to the storing of
-                        cookies on your device to enhance site navigation,
-                        improve your experience, and assist in our marketing
-                        efforts. To learn more, read our Privacy Policy.
-                    </p>
-
-                    <div className="mt-8 flex gap-3 justify-end">
-                        <button
-                            className="text-white bg-[#7147E8] rounded-3xl py-2 px-8 text-sm"
-                            onClick={() => handleButtonClick('Accept')}
-                        >
-                            Accept
-                        </button>
-                        <button
-                            className="text-[#7147E8] bg-white border border-[#7147E8] rounded-3xl py-2 px-8 text-sm"
-                            onClick={() => handleButtonClick('Decline')}
-                        >
-                            Decline
-                        </button>
-                    </div>
-                </div>
-            )}
         <Header onClick={handleTryOpenMetadataClick} />
         <Achievement />
         <Services />
